+++ src/backend/index.ts
@@ -15,6 +15,7 @@ import notificationsRouter from './src/routes/notifications.routes';
 import designReviewsRouter from './src/routes/design-reviews.routes';
 import workPackageTemplatesRouter from './src/routes/work-package-templates.routes';
 import carsRouter from './src/routes/cars.routes';
+import organizationRouter from './src/routes/organizations.routes';
 
 const app = express();
 const port = process.env.PORT || 3001;
@@ -60,6 +61,7 @@ app.use('/design-reviews', designReviewsRouter);
 app.use('/notifications', notificationsRouter);
 app.use('/templates', workPackageTemplatesRouter);
 app.use('/cars', carsRouter);
+app.use('/organizations', organizationRouter);
 app.use('/', (_req, res) => {
   res.json('Welcome to FinishLine');
 });
